demote CUDA and Term to weak dependencies
