+++ Project.toml
@@ -1,11 +1,10 @@
 name = "XGBoost"
 uuid = "009559a3-9522-5dbb-924b-0b6ed2b22bb9"
-version = "2.2.5"
+version = "2.2.6"
 
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
 CEnum = "fa961155-64e5-5f13-b03f-caf6b980ea82"
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
@@ -13,9 +12,16 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SparseMatricesCSR = "a0a7dd2c-ebf4-11e9-1f05-cf50bc540ca1"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
-Term = "22787eb5-b846-44ae-b979-8e399b8463ab"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 XGBoost_jll = "a5c6f535-4255-5ca2-a466-0e519f119c46"
+
+[weakdeps]
+Term = "22787eb5-b846-44ae-b979-8e399b8463ab"
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+
+[extensions]
+CUDAExt = "CUDA"
+TermExt = "Term"
 
 [compat]
 AbstractTrees = "0.4"
@@ -32,6 +38,10 @@ julia = "1.6"
 [extras]
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+# cuda is needed just to check if gpu exists
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+# this is needed for backward compat with pre-1.9
+Term = "22787eb5-b846-44ae-b979-8e399b8463ab"
 
 [targets]
-test = ["Random", "Test"]
+test = ["Random", "Test", "CUDA", "Term"]
