+++ src/dmatrix.jl
@@ -184,25 +184,6 @@ function _dmatrix(x::AbstractMatrix{T}; missing_value::Float32=NaN32, kw...) whe
     DMatrix(o[]; kw...)
 end
 
-# sadly we have to copy CuArray because of incompatible column convention
-function _transposed_cuda_dmatrix(x::CuArray{T}; missing_value::Float32=NaN32, kw...) where {T<:Real}
-    o = Ref{DMatrixHandle}()
-    cfg = "{\"missing\": $missing_value}"
-    GC.@preserve x begin
-        info = numpy_json_info(x)
-        xgbcall(XGDMatrixCreateFromCudaArrayInterface, info, cfg, o)
-    end
-    DMatrix(o[]; is_gpu=true, kw...)
-end
-
-DMatrix(x::Transpose{T,<:CuArray}; kw...) where {T<:Real} = _transposed_cuda_dmatrix(parent(x); kw...)
-DMatrix(x::Adjoint{T,<:CuArray}; kw...) where {T<:Real} = _transposed_cuda_dmatrix(parent(x); kw...)
-
-function DMatrix(x::CuArray; kw...)
-    x′ = CuArray(transpose(x))
-    _transposed_cuda_dmatrix(x′; kw...)
-end
-
 function DMatrix(x::AbstractMatrix{T}; kw...) where {T<:Real}
     # sadly, this copying is unavoidable
     _dmatrix(convert(Matrix{Float32}, transpose(x)); kw...)
@@ -270,21 +251,12 @@ DMatrix(Xy::Tuple; kw...) = DMatrix(Xy[1], Xy[2]; kw...)
 
 DMatrix(dm::DMatrix) = dm
 
-function _check_gpu_table(tbl)
-    cols = Tables.Columns(tbl)
-    isgpu = all(x -> x isa CuArray, cols)
-    (isgpu, cols)
-end
+# this is done this way so that we can simply add methods in the extension
+_check_gpu_col(::AbstractArray) = false
+_check_gpu_table(cols) = all(_check_gpu_col, cols)
 
-function _dmatrix_gpu_table(cols::Tables.Columns; missing_value::Float32=NaN32, kw...)
-    o = Ref{DMatrixHandle}()
-    cfg = "{\"missing\": $missing_value}"
-    GC.@preserve cols begin
-        infos = numpy_json_infos(cols)
-        xgbcall(XGDMatrixCreateFromCudaColumnar, infos, cfg, o)
-    end
-    DMatrix(o[]; is_gpu=true, kw...)
-end
+# extended by CUDAExt
+function _dmatrix_gpu_table end
 
 function DMatrix(tbl;
                  feature_names::AbstractVector{<:AbstractString}=collect(string.(Tables.columnnames(tbl))),
@@ -293,8 +265,9 @@ function DMatrix(tbl;
     if !Tables.istable(tbl)
         throw(ArgumentError("DMatrix requires either an AbstractMatrix or table satisfying the Tables.jl interface"))
     end
-    (isgpu, cols) = _check_gpu_table(tbl)
-    if isgpu
+    cols = Tables.Columns(tbl)
+    if _check_gpu_table(cols)
+        # only reachable if CUDAExt is loaded
         _dmatrix_gpu_table(cols; feature_names, kw...)
     else
         DMatrix(Tables.matrix(tbl); feature_names, kw...)
