+++ test/runtests.jl
@@ -1,4 +1,5 @@
 using XGBoost
+using Term
 using CUDA: has_cuda, cu
 using Random, SparseArrays
 using Test
@@ -160,7 +161,7 @@ end
     @test tbl.feature[1] == 29
     @test XGBoost.Tables.columnnames(tbl) == (:feature, :gain, :weight, :cover, :total_gain, :total_cover)
 
-    @test typeof(importancereport(bst)) <: XGBoost.Term.Tables.Table
+    @test typeof(importancereport(bst)) <: Term.Tables.Table
 end
 
 @testset "Booster Save/Load/Serialize" begin
